+++ README.md
@@ -7,12 +7,27 @@
 
 This plugin adds an event listener to log emails sent on `mail_logs` database table. It also adds a Filament resource to view the mail logs.
 
+## Version Compatibility
+
+ Filament | Filament MailLog
+:---------|:---------------------------
+ 3.x      | 1.x
+ 4.x      | 2.x
+
 ## Installation
 
 You can install the package via composer:
 
+### For Filament 3
+
 ```bash
-composer require tapp/filament-maillog
+composer require tapp/filament-maillog:"^1.0"
+```
+
+### For Filament 4
+
+```bash
+composer require tapp/filament-maillog:"^2.0"
 ```
 
 You can publish and run the migrations with:
